Contrib: add decomposition submodule
