+++ tensorly/contrib/__init__.py
@@ -2,4 +2,3 @@
 
     Allows to add quickly and test new functions for which the API is not necessarily fixed
 """
-from .mps_decomposition_cross import matrix_product_state_cross
