+++ tensorly/contrib/decomposition/__init__.py
@@ -1,0 +1,7 @@
+"""
+The :mod:`tensorly.contrib.decomposition` module includes utilities for performing
+tensor decomposition for which the API is still evolving.                                                                                               
+"""
+
+from .mps_decomposition_cross import matrix_product_state_cross
+
